requirements.txt
requests==2.32.3
yt_dlp==2025.6.9
supabase==2.15.3
python-dotenv==1.1.0